LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/fill-and-stroke-styles/2d.pattern.transform.infinity-expected.txt

2d.pattern.transform.infinity
Actual output:
Expected output:


FAIL Canvas test: 2d.pattern.transform.infinity assert_equals: Green channel of the pixel at (1, 1) expected 255 but got 0
